llm/backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
langchain==0.1.0
langchain-openai==0.0.5
python-dotenv==1.0.0
openai==1.12.0